+++ src/asm/main.rs
@@ -52,7 +52,7 @@ pub struct RuntimeOptions {
     binary_digits: [char; 2],
     gfx_chars: [char; 4],
     pad_byte: u8,
-    q_precision: usize,
+    q_precision: u8,
     recursion_depth: usize,
     // TODO(perf): use some bitfield(s) instead?
     warnings: [WarningState; NB_WARNINGS],
